feat(v0.2.0): update checker, bug fixes, memory leak fixes
- New: GitHub releases-based update checker with one-click installer
- New: Check for Updates button in app header
- Fix: extractor.all_quality_passed unbounded growth - now bounded heap (top-K)
- Fix: DiversityFilter._accepted unbounded growth - trim to max_compare
- Fix: ColorDiversityFilter._fingerprints unbounded growth - same
- Fix: completeness.compute_completeness_score uint8 overflow on >16MP frames
- Fix: resize.generate_buckets validates min_bucket/max_bucket alignment
- Fix: resize.cover_resize_and_crop guards against zero-dim images
- Fix: app.py uses tkinter.messagebox.showerror (not CTkInputDialog) for errors
- Fix: app.py refreshes parameter labels on language toggle
- Fix: build_exe.py declares all hidden imports for cv2/numpy/scenedetect/PIL
- Fix: filedialog uses tkinter directly (more reliable than ctk re-export)
- New i18n keys: select_video_folder, select_output_folder, update_*

+++ src/vid2dataset/completeness.py
@@ -40,14 +40,15 @@ def compute_completeness_score(frame_bgr: np.ndarray, *, border_ratio: float = 0
 
     center_mask = 1 - border_mask
 
-    border_pixels = border_mask.sum()
-    center_pixels = center_mask.sum()
+    border_pixels = int(border_mask.sum())
+    center_pixels = int(center_mask.sum())
 
     if border_pixels == 0 or center_pixels == 0:
         return 1.0
 
-    border_edge_density = (edges * border_mask).sum() / (border_pixels * 255)
-    center_edge_density = (edges * center_mask).sum() / (center_pixels * 255)
+    # Cast to int64 to prevent uint8 overflow on large images.
+    border_edge_density = int((edges.astype(np.int64) * border_mask).sum()) / (border_pixels * 255)
+    center_edge_density = int((edges.astype(np.int64) * center_mask).sum()) / (center_pixels * 255)
 
     # If border has more edge activity relative to center, subject is cut off
     if center_edge_density < 0.001:
